change d matrix to float

+++ aligner.h
@@ -8,7 +8,7 @@ void aligner(bwt_t *const bwt, int len, ubyte_t *seq, bwtint_t level, hash_eleme
 
 void
 create_cigar(aryana_args *options, hash_element *best, char *cigar, int len, const ubyte_t *seq, const ubyte_t *qual,
-             uint64_t seq_len, int **d, char **arr, char *tmp_cigar, penalty_t *penalty, uint64_t *reference,
+             uint64_t seq_len, float **d, char **arr, char *tmp_cigar, penalty_t *penalty, uint64_t *reference,
              ignore_mismatch_t ignore);
 
 void showerr(int len, const ubyte_t *seq);
